+++ packages/app-core/src/JBrowseModel/index.ts
@@ -1,0 +1,160 @@
+import PluginManager from '@jbrowse/core/PluginManager'
+import { BaseAssemblyConfigSchema } from '@jbrowse/core/assemblyManager'
+import { JBrowseConfigF } from '../JBrowseConfig'
+import { cast, getParent } from 'mobx-state-tree'
+import RpcManager from '@jbrowse/core/rpc/RpcManager'
+import {
+  AnyConfigurationModel,
+  readConfObject,
+} from '@jbrowse/core/configuration'
+import { PluginDefinition } from '@jbrowse/core/PluginLoader'
+
+export function JBrowseModelF(
+  pluginManager: PluginManager,
+  assemblyConfigSchemasType: BaseAssemblyConfigSchema,
+) {
+  return JBrowseConfigF(pluginManager, assemblyConfigSchemasType)
+    .views(self => ({
+      /**
+       * #getter
+       */
+      get assemblyNames(): string[] {
+        return self.assemblies.map(assembly => readConfObject(assembly, 'name'))
+      },
+      /**
+       * #getter
+       */
+      get rpcManager(): RpcManager {
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        return getParent<any>(self).rpcManager
+      },
+    }))
+    .actions(self => ({
+      afterCreate() {
+        const seen = [] as string[]
+        self.assemblyNames.forEach(assemblyName => {
+          if (!assemblyName) {
+            throw new Error('Encountered an assembly with no "name" defined')
+          }
+          if (seen.includes(assemblyName)) {
+            throw new Error(
+              `Found two assemblies with the same name: ${assemblyName}`,
+            )
+          } else {
+            seen.push(assemblyName)
+          }
+        })
+      },
+      /**
+       * #action
+       */
+      addAssemblyConf(conf: AnyConfigurationModel) {
+        const { name } = conf
+        if (!name) {
+          throw new Error('Can\'t add assembly with no "name"')
+        }
+        if (self.assemblyNames.includes(name)) {
+          throw new Error(
+            `Can't add assembly with name "${name}", an assembly with that name already exists`,
+          )
+        }
+        const length = self.assemblies.push({
+          ...conf,
+          sequence: {
+            type: 'ReferenceSequenceTrack',
+            trackId: `${name}-${Date.now()}`,
+            ...conf.sequence,
+          },
+        })
+        return self.assemblies[length - 1]
+      },
+      /**
+       * #action
+       */
+      removeAssemblyConf(assemblyName: string) {
+        const toRemove = self.assemblies.find(a => a.name === assemblyName)
+        if (toRemove) {
+          self.assemblies.remove(toRemove)
+        }
+      },
+      /**
+       * #action
+       */
+      addTrackConf(trackConf: AnyConfigurationModel) {
+        const { type } = trackConf
+        if (!type) {
+          throw new Error(`unknown track type ${type}`)
+        }
+        const length = self.tracks.push(trackConf)
+        return self.tracks[length - 1]
+      },
+      /**
+       * #action
+       */
+      addConnectionConf(connectionConf: AnyConfigurationModel) {
+        const { type } = connectionConf
+        if (!type) {
+          throw new Error(`unknown connection type ${type}`)
+        }
+        const length = self.connections.push(connectionConf)
+        return self.connections[length - 1]
+      },
+      /**
+       * #action
+       */
+      deleteConnectionConf(configuration: AnyConfigurationModel) {
+        const elt = self.connections.find(conn => conn.id === configuration.id)
+        return self.connections.remove(elt)
+      },
+      /**
+       * #action
+       */
+      deleteTrackConf(trackConf: AnyConfigurationModel) {
+        const elt = self.tracks.find(t => t.trackId === trackConf.trackId)
+        return self.tracks.remove(elt)
+      },
+      /**
+       * #action
+       */
+      addPlugin(pluginDefinition: PluginDefinition) {
+        self.plugins.push(pluginDefinition)
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        const rootModel = getParent<any>(self)
+        rootModel.setPluginsUpdated(true)
+      },
+      /**
+       * #action
+       */
+      removePlugin(pluginDefinition: PluginDefinition) {
+        self.plugins = cast(
+          self.plugins.filter(
+            plugin =>
+              plugin.url !== pluginDefinition.url ||
+              plugin.umdUrl !== pluginDefinition.umdUrl ||
+              plugin.cjsUrl !== pluginDefinition.cjsUrl ||
+              plugin.esmUrl !== pluginDefinition.esmUrl,
+          ),
+        )
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        getParent<any>(self).setPluginsUpdated(true)
+      },
+      /**
+       * #action
+       */
+      addInternetAccountConf(internetAccountConf: AnyConfigurationModel) {
+        const { type } = internetAccountConf
+        if (!type) {
+          throw new Error(`unknown internetAccount type ${type}`)
+        }
+        const length = self.internetAccounts.push(internetAccountConf)
+        return self.internetAccounts[length - 1]
+      },
+      /**
+       * #action
+       */
+      deleteInternetAccountConf(configuration: AnyConfigurationModel) {
+        const elt = self.internetAccounts.find(a => a.id === configuration.id)
+        return self.internetAccounts.remove(elt)
+      },
+    }))
+}
